user login working on development server
